Cleaning up files.
