--- class/NominationEmail.php
@@ -1,207 +1,0 @@
-<?php
-namespace nomination;
-use exception\DatabaseException;
-/**
- * NominationEmail
- *
- * Handles sending emails to various people for the nomination module.
- *
- * @author Daniel West <dwest at tux dot appstate dot edu>
- * @author Jeremy Booker
- * @author Chris Detsch
- * @package nomination
- */
-
-\PHPWS_Core::initCoreClass('Mail.php');
-
-// Abbreviated NominationActor names
-// Used in DB
-define('SHORT_Nominator', 'NMR');
-define('SHORT_Nominee',   'NME');
-define('SHORT_Reference', 'REF');
-
-// Message types
-define('NEWNOM', 'New Nomination');
-define('NEWREF', 'New Reference');
-define('UPDNOM', 'Updated nomination');
-define('REFUPL', 'Reference document upload');
-define('NOMDEL', 'Removal request approved');
-define('REFDEL', 'Removal request approved');
-
-define('ALLNOM', 'All nominators');
-define('NOMCPL', 'Nominators with complete nomination');
-define('NOMINC', 'Nominators with incomplete nomination');
-define('REFNON', 'References that need to upload');
-define('NOMINE', 'Nominees with complete nominations');
-
-
-
-abstract class NominationEmail {
-
-    public $from;
-    public $list;
-    public $subject;
-    public $message;
-    public $messageType;
-
-    public function __construct($subject, $message, $msgType)
-    {
-        $this->subject     = $subject;
-        $this->message     = $message;
-        $this->messageType = $msgType;
-        $this->from        = \PHPWS_Settings::get('nomination', 'email_from_address');
-    }
-
-    public function sendTo($recipientEmail)
-    {
-
-        $mail = new \PHPWS_Mail;
-        $mail->sendIndividually(true);
-
-        $mail->addSendTo($recipientEmail);
-
-        $mail->setFrom($this->from);
-        $mail->setSubject($this->subject);
-        $mail->setMessageBody($this->message);
-
-        if(!EMAIL_TEST_FLAG){
-            $mail->send();
-        }
-
-    }
-
-    // Build NominationEmail from EmailMessage and send it.
-    // This may be a deprecated method
-    public static function sendMessageObj(EmailMessage $msg)
-    {
-        switch($msg->receiver_type)
-        {
-            case SHORT_Reference:
-                \PHPWS_Core::initModClass('nomination', 'Reference.php');
-                $db = new \PHPWS_DB('nomination_reference');
-                $factory = new ReferenceFactory();
-                $obj = $factory->getReferenceById($msg->nominee_id);
-                break;
-            case SHORT_Nominator:
-                \PHPWS_Core::initModClass('nomination', 'Nominator.php');
-                $db = new \PHPWS_DB('nomination_nominator');
-                $obj = new Nominator();
-                break;
-            case SHORT_Nominee:
-                \PHPWS_Core::initModClass('nomination', 'Nominee.php');
-                $db = new \PHPWS_DB('nomination_nominee');
-                $obj = new Nominee();
-                break;
-        }
-
-        // Get the email address.
-        $db->addWhere('id', $msg->receiver_id);
-        $result = $db->loadObject($obj);
-
-
-        $obj = array($obj);
-        $nominationEmail = new NominationEmail($obj, $msg->subject, $msg->message, $msg->message_type);
-        $nominationEmail->send();
-    }
-
-    public function logEmail(Nomination $nomination, $recipientEmail, $recipientId, $receiverType)
-    {
-        // This is all kinds of messed up. Just write it to a log file for now...
-
-
-
-        // Log the message to a text file
-        $fd = fopen(PHPWS_SOURCE_DIR . 'logs/email.log',"a");
-        fprintf($fd, "=======================\n");
-
-
-        fprintf($fd, "To: %s\n", $recipientEmail);
-
-        fprintf($fd, "From: %s\n", $this->from);
-        fprintf($fd, "Subject: %s\n", $this->subject);
-        fprintf($fd, "Content: \n");
-        fprintf($fd, "%s\n\n", $this->message);
-
-        fclose($fd);
-
-
-
-
-        $messageLog = new EmailLog($nomination->getId(), $this->message,
-        $this->messageType, $this->subject, $recipientId, $receiverType, time());
-
-
-        /*
-        $messageLog->setNomineeId($nomination->getId();
-
-        $messageLog->setMessage($this->message);
-        $messageLog->setMessageType($this->messageType);
-        $messageLog->setSubject($this->subject);
-        $messageLog->setReceiverId($recipientId);
-        $messageLog->setReceiverType($receiverType);
-        $messageLog->setSentOn($now);
-        */
-
-        EmailLogFactory::save($messageLog);
-
-
-    }
-
-    public static function getLongMessageType($type)
-    {
-        switch($type){
-            case 'NEWNOM':
-                return NEWNOM;
-            case 'UPDNOM':
-                return UPDNOM;
-            case 'REFUPL':
-                return REFUPL;
-            case 'ALLNOM':
-                return ALLNOM;
-            case 'NOMCPL':
-                return NOMCPL;
-            case 'NOMINC':
-                return NOMINC;
-            case 'REFNON':
-                return REFNON;
-            case 'NOMINE':
-                return NOMINE;
-            default:
-                return null;
-        }
-    }
-
-    // TODO: Is there a way to build a constant's name with string
-    public static function getAbbrevName($class)
-    {
-        switch($class){
-            case 'Nominator':
-                return SHORT_Nominator;
-            case 'Reference':
-                return SHORT_Reference;
-            case 'Nominee':
-                return SHORT_Nominee;
-        }
-    }
-
-    public static function getLists()
-    {
-        //if you change anything about this array update the below function
-        //yes it's hackish but we're "sure" that there will only be 5 lists...
-        $lists = array();
-
-        $lists['ALLNOM'] = ALLNOM;
-        $lists['NOMCPL'] = NOMCPL;
-        $lists['NOMINC'] = NOMINC;
-        $lists['REFNON'] = REFNON;
-        $lists['NOMINE'] = NOMINE;
-
-        // These must be here for EmailLog to properly work
-        // These later become unset in the SendEmail class.
-        $lists['NEWNOM'] = NEWNOM;
-        $lists['NEWREF'] = NEWREF;
-
-        return $lists;
-    }
-
-}
