+++ class/command/EditEmail.php
@@ -12,7 +12,6 @@ namespace nomination\command;
 
 use \nomination\Command;
 use \nomination\Context;
-use \nomination\NominationEmail;
 use \nomination\UserStatus;
 
 class EditEmail extends Command {
