+++ inc/defines.php
@@ -17,7 +17,7 @@ define('NOMINATION_FACULTY_MEMBER_TEXT', 'Faculty Member');
 define('NOMINATION_EMPLOYEE_TEXT', 'Employee of ASU');
 
 define('NOMINATION_EMAIL_DOMAIN', 'appstate.edu');
-define('EMAIL_TEST_FLAG', FALSE);
+//define('EMAIL_TEST_FLAG', FALSE);
 
 define('NOMINATOR', 'NTR');
 define('REFERENCE', 'REF');
